# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "配置平台 HTTP接口文档",
        "title": "配置平台",
        "contact": {
            "name": "李明",
            "email": "2572597150@qq.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/auth/refreshToken": {
            "get": {
                "tags": [
                    "鉴权"
                ],
                "summary": "刷新token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LoginResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.AuthError"
                        }
                    }
                }
            }
        },
        "/captcha/confirm": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "鉴权"
                ],
                "summary": "校验图片验证码",
                "parameters": [
                    {
                        "description": "校验图片验证码",
                        "name": "ConfirmCaptcha",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ConfirmCaptcha"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CaptchaToken"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/services.WebResponse"
                        }
                    }
                }
            }
        },
...
        "/email/confirm": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "鉴权"
                ],
                "summary": "验证邮箱验证码",
                "parameters": [
                    {
                        "description": "图片验证码返回的token",
                        "name": "ConfirmEmailCode",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EmailCode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/services.WebResponse"
                        }
                    }
                }
            }
        },
        "/email/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "鉴权"
                ],
                "summary": "发送邮箱验证码",
                "parameters": [
                    {
                        "description": "图片验证码返回的token",
                        "name": "SendEmailCode",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SendEmailCode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/services.WebResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AuthError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.CaptchaToken": {
            "type": "object",
            "properties": {
                "captcha_token": {
                    "description": "图片验证码验证成功返回的token",
                    "type": "string"
                }
            }
        },
        "model.ConfirmCaptcha": {
            "type": "object",
            "required": [
                "captcha_id",
                "captcha_solution"
            ],
            "properties": {
                "captcha_id": {
                    "description": "验证码id",
                    "type": "string"
                },
                "captcha_solution": {
                    "description": "验证码的值",
                    "type": "string"
                }
            }
        },
...
        "model.EmailCode": {
            "type": "object",
            "required": [
                "email_code"
            ],
            "properties": {
                "email_code": {
                    "description": "邮箱验证码",
                    "type": "string"
                }
            }
        },
...
        "model.LoginResp": {
            "type": "object",
            "properties": {
                "expire": {
                    "description": "过期时间",
                    "type": "string"
                },
                "token": {
                    "description": "token",
                    "type": "string"
                }
            }
        },
...
        "model.SendEmailCode": {
            "type": "object",
            "required": [
                "captcha_token",
                "email_address"
            ],
            "properties": {
                "captcha_token": {
                    "description": "图片验证码验证成功返回的token",
                    "type": "string"
                },
                "email_address": {
                    "description": "邮箱地址",
                    "type": "string"
                }
            }
        },
        "services.WebResponse": {
            "type": "object",
            "properties": {
                "errcode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errmsg": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        }
    }
}